pyproject.toml
[project]
name = "dummy-ctrl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gym-aloha-dummy",
    "ipykernel>=6.29.5",
    "mediapipe>=0.10.14",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pynput>=1.8.1",
    "pyusb>=1.3.1",
    "telemoma",
]

[tool.uv.sources]
telemoma = { path = "third_party/telemoma_aloha", editable = true }
gym-aloha-dummy = { path = "third_party/gym-aloha", editable = true }
